+++ spec/spec_helper.rb
@@ -3,22 +3,36 @@
 require 'simplecov'
 SimpleCov.start
 
-module Legion
-  module Logging
-    def self.debug(_msg = nil); end
-    def self.info(_msg = nil); end
-    def self.warn(_msg = nil); end
-    def self.error(_msg = nil); end
-  end
+require 'legion/logging'
+require 'legion/settings'
+require 'legion/json'
+require 'legion/data'
+require 'legion/cache'
+require 'legion/crypt'
+require 'legion/transport'
 
+module Legion
   module Extensions
     module Core
+    end
+
+    module Helpers
+      module Lex
+        def self.included(base)
+          base
+        end
+      end
+    end
+
+    module Actors
+      class Every; end # rubocop:disable Lint/EmptyClass
     end
   end
 end
 
 $LOADED_FEATURES << 'legionio.rb'
 $LOADED_FEATURES << 'legion/extensions/core.rb'
+$LOADED_FEATURES << 'legion/extensions/actors/every'
 
 lib = File.expand_path('../lib', __dir__)
 $LOAD_PATH.unshift(lib) unless $LOAD_PATH.include?(lib)